requirements.txt
asgiref==3.5.0
astroid==2.9.3
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pylint==2.12.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
whitenoise==6.0.0
wrapt==1.13.3
